+++ relict/__init__.py
@@ -5,8 +5,9 @@
 from .cli import main
 from .home import documents_dir, drives_dir, home, set_home
 from .interaction import GuestExec
-from .interaction_agentless import (AgentlessGuestExec, screen_text,
-                                    send_keys, send_text, wait_screen)
+from .interaction_agentless import (AgentlessGuestExec,
+                                    cursor_menu_select, screen_text,
+                                    send_keys, send_text, wait_text)
 from .lifecycle import (Qmp, create_hdd_image, find_qemu, find_qemu_img,
                         stop)
 from .machine import Machine, screenshot
@@ -21,6 +22,7 @@ __all__ = [
     "AgentlessGuestExec",
     "GuestExec",
     "create_hdd_image",
+    "cursor_menu_select",
     "documents_dir",
     "drives_dir",
     "find_qemu",
@@ -36,5 +38,5 @@ __all__ = [
     "set_home",
     "start",
     "stop",
-    "wait_screen",
+    "wait_text",
 ]